+++ minix/drivers/storage/fbd/rule.h
@@ -11,7 +11,7 @@ extern int rule_find(u64_t pos, size_t size, int flag);
 extern void rule_pre_hook(iovec_t *iov, unsigned *count, size_t *size,
 	u64_t *pos);
 extern void rule_io_hook(char *buf, size_t size, u64_t pos, int flag);
-extern void rule_post_hook(size_t osize, ssize_t *result);
+extern void rule_post_hook(size_t osize, int *result);
 
 #define PRE_HOOK	0x1
 #define IO_HOOK		0x2
