+++ minix/drivers/storage/vnd/vnd.c
@@ -28,7 +28,7 @@ static unsigned int instance;
 
 static int vnd_open(devminor_t, int);
 static int vnd_close(devminor_t);
-static ssize_t vnd_transfer(devminor_t, int, u64_t, endpoint_t, iovec_t *,
+static int vnd_transfer(devminor_t, int, u64_t, endpoint_t, iovec_t *,
 	unsigned int, int);
 static int vnd_ioctl(devminor_t, unsigned long, endpoint_t, cp_grant_id_t,
 	endpoint_t);
@@ -140,7 +140,7 @@ static int vnd_copy(iovec_s_t *iov, size_t iov_off, size_t bytes, endpoint_t end
 		chunk = MIN(bytes - off, iov->iov_size - iov_off);
 
 		if (endpt == SELF) {
-			ptr = (char *)(uintptr_t) iov->iov_grant + iov_off;
+			ptr = (char *) iov->iov_grant + iov_off;
 
 			if (do_write)
 				memcpy(&state.buf[off], ptr, chunk);
@@ -204,7 +204,7 @@ vnd_advance(iovec_s_t *iov, size_t *iov_offp, size_t bytes)
 /*
  * Perform data transfer on the selected device.
  */
-static ssize_t vnd_transfer(devminor_t minor, int do_write, u64_t position,
+static int vnd_transfer(devminor_t minor, int do_write, u64_t position,
 	endpoint_t endpt, iovec_t *iovt, unsigned int nr_req, int flags)
 {
 	struct device *dv;
@@ -265,7 +265,7 @@ static ssize_t vnd_transfer(devminor_t minor, int do_write, u64_t position,
 
 		/* Copy the data for this chunk from or to the caller. */
 		if ((r = vnd_copy(iov, iov_off, chunk, endpt, do_write)) < 0) {
-			printf("VND%u: data copy failed (%zd)\n", instance, r);
+			printf("VND%u: data copy failed (%d)\n", instance, r);
 			return r;
 		}
 
@@ -277,7 +277,7 @@ static ssize_t vnd_transfer(devminor_t minor, int do_write, u64_t position,
 			if (r <= 0) {
 				if (r < 0)
 					r = -errno;
-				printf("VND%u: pwrite failed (%zd)\n", instance,
+				printf("VND%u: pwrite failed (%d)\n", instance,
 				    r);
 				return (r < 0) ? r : EIO;
 			}
